SKALE-1653 new node_data location, restructure config dirs on host, refactor confiigs module
