+++ cli/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.5.0'
+__version__ = '0.6.0'
 
 if __name__ == "__main__":
     print(__version__)
