+++ cli/node.py
@@ -30,8 +30,7 @@ from core.node import create_node, init, purge, update
 from core.host import install_host_dependencies
 from core.helper import (abort_if_false, local_only,
                          login_required, safe_load_texts)
-from core.config import CONFIG_FILEPATH, DEFAULT_DB_USER, DEFAULT_DB_PORT
-from configs.node import DEFAULT_NODE_BASE_PORT
+from configs import CONFIG_FILEPATH, DEFAULT_DB_USER, DEFAULT_DB_PORT, DEFAULT_NODE_BASE_PORT
 
 config = ReadSettings(CONFIG_FILEPATH)
 TEXTS = safe_load_texts()
@@ -201,11 +200,6 @@ def register_node(name, ip, port):
     help='URL to IMA contracts ABI and addresses'
 )
 @click.option(
-    '--dkg-url',
-    prompt="Enter URL to DKG contracts ABI and addresses",
-    help='URL to DKG contracts ABI and addresses'
-)
-@click.option(
     '--filebeat-url',
     prompt="Enter URL to the Filebeat log server",
     help='URL to the Filebeat log server'
@@ -217,16 +211,14 @@ def register_node(name, ip, port):
 @local_only
 def init_node(ima_endpoint, install_deps, stream, github_token, docker_username, docker_password,
               endpoint, db_user, db_password, db_root_password, db_port, disk_mountpoint,
-              manager_url, ima_url, dkg_url, filebeat_url, test_mode):
+              manager_url, ima_url, filebeat_url, test_mode):
     if install_deps:
         install_host_dependencies()
     if not db_root_password:
         db_root_password = db_password
-
-    git_branch = stream
-    init(ima_endpoint, git_branch, github_token, docker_username, docker_password, endpoint,
+    init(ima_endpoint, stream, github_token, docker_username, docker_password, endpoint,
          db_user, db_password, db_root_password, db_port, disk_mountpoint, manager_url, ima_url,
-         dkg_url, filebeat_url, test_mode)
+         filebeat_url, test_mode)
 
 
 @node.command('purge', help="Uninstall SKALE node software from the machine")
@@ -313,20 +305,14 @@ def purge_node():
     help='URL to IMA contracts ABI and addresses'
 )
 @click.option(
-    '--dkg-url',
-    prompt="Enter URL to DKG contracts ABI and addresses",
-    help='URL to DKG contracts ABI and addresses'
-)
-@click.option(
     '--filebeat-url',
     prompt="Enter URL to the Filebeat log server",
     help='URL to the Filebeat log server'
 )
 @local_only
 def update_node(ima_endpoint, github_token, docker_username, docker_password, endpoint, db_user,
-                db_password, db_root_password, db_port, manager_url, ima_url, dkg_url,
-                filebeat_url):
+                db_password, db_root_password, db_port, manager_url, ima_url, filebeat_url):
     if not db_root_password:
         db_root_password = db_password
     update(ima_endpoint, github_token, docker_username, docker_password, endpoint, db_user,
-           db_password, db_root_password, db_port, manager_url, ima_url, dkg_url, filebeat_url)
+           db_password, db_root_password, db_port, manager_url, ima_url, filebeat_url)
