--- configs/node.py
@@ -1,30 +1,0 @@
-#   -*- coding: utf-8 -*-
-#
-#   This file is part of skale-node-cli
-#
-#   Copyright (C) 2019 SKALE Labs
-#
-#   This program is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU Affero General Public License as published by
-#   the Free Software Foundation, either version 3 of the License, or
-#   (at your option) any later version.
-#
-#   This program is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty of
-#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#   GNU General Public License for more details.
-#
-#   You should have received a copy of the GNU Affero General Public License
-#   along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-import os
-from pathlib import Path
-
-SKALE_VOLUME_PATH = '/skale_vol'
-NODE_DATA_PATH = '/skale_node_data'
-HOME_DIR = str(Path.home())
-
-LOCAL_WALLET_FILENAME = 'local_wallet.json'
-LOCAL_WALLET_FILEPATH = os.path.join(NODE_DATA_PATH, LOCAL_WALLET_FILENAME)
-
-DEFAULT_NODE_BASE_PORT = 10000
