+++ configs/routes.py
@@ -1,0 +1,21 @@
+ROUTES = {
+    'login': '/login',
+    'logout': '/logout',
+    'register': '/join',
+    'node_info': '/node-info',
+    'node_about': '/about-node',
+    'create_node': '/create-node',
+    'test_host': '/test-host',
+
+    'wallet_info': '/load-wallet',
+    'validators_info': '/validators-info',
+
+    'schains_containers': '/containers/schains/list',
+    'node_schains': '/schains/list',
+    'schain_config': '/schain-config',
+    'skale_containers': '/containers/list',
+
+    'logs_dump': '/logs/dump',
+    'first-bounties': '/first-bounties',
+    'last-bounties': '/last-bounties'
+}
