+++ core/validators.py
@@ -17,13 +17,13 @@
 #   You should have received a copy of the GNU Affero General Public License
 #   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from core.config import URLS
+from configs import ROUTES
 from core.helper import get_node_creds, construct_url, get_request
 
 
 def get_validators_info(config, format):
     host, cookies = get_node_creds(config)
-    url = construct_url(host, URLS['validators_info'])
+    url = construct_url(host, ROUTES['validators_info'])
 
     response = get_request(url, cookies)
     if response is None:
